rethinink behavior composition
